continued implementation of receive and ship inventory

+++ app/src/main/java/org/helpingkidsroundfirst/hkrf/data/InventoryDbHelper.java
@@ -9,6 +9,8 @@ import org.helpingkidsroundfirst.hkrf.data.InventoryContract.CategoryEntry;
 import org.helpingkidsroundfirst.hkrf.data.InventoryContract.CurrentInventoryEntry;
 import org.helpingkidsroundfirst.hkrf.data.InventoryContract.ItemEntry;
 import org.helpingkidsroundfirst.hkrf.data.InventoryContract.PastInventoryEntry;
+import org.helpingkidsroundfirst.hkrf.data.InventoryContract.ReceiveInventoryEntry;
+import org.helpingkidsroundfirst.hkrf.data.InventoryContract.ShipInventoryEntry;
 
 /**
  * Manages the local database
@@ -19,7 +21,7 @@ public class InventoryDbHelper extends SQLiteOpenHelper {
 
     static final String DATABASE_NAME = "inventory.db";
     //change when database changes
-    private static final int DATABASE_VERSION = 6;
+    private static final int DATABASE_VERSION = 7;
 
     public InventoryDbHelper(Context context) {
         super(context, DATABASE_NAME, null, DATABASE_VERSION);
@@ -62,6 +64,20 @@ public class InventoryDbHelper extends SQLiteOpenHelper {
                 //set up foreign key for item
                 "FOREIGN KEY (" + CurrentInventoryEntry.COLUMN_ITEM_KEY + ") REFERENCES " +
                 ItemEntry.TABLE_NAME + " (" + ItemEntry._ID + "));";
+
+        // create receive inventory table
+        final String SQL_CREATE_RECEIVE_INVENTORY_TABLE = "CREATE TABLE " +
+                ReceiveInventoryEntry.TABLE_NAME + " (" +
+                ReceiveInventoryEntry._ID + " INTEGER PRIMARY KEY, " +
+                ReceiveInventoryEntry.COLUMN_ITEM_KEY + " INTEGER, " +
+                ReceiveInventoryEntry.COLUMN_QTY + " INTEGER, " +
+                ReceiveInventoryEntry.COLUMN_DATE_RECEIVED + " INTEGER, " +
+                ReceiveInventoryEntry.COLUMN_DONOR + " TEXT, " +
+                ReceiveInventoryEntry.COLUMN_WAREHOUSE + " TEXT, " +
+
+                //set up foreign key for item
+                "FOREIGN KEY (" + CurrentInventoryEntry.COLUMN_ITEM_KEY + ") REFERENCES " +
+                ItemEntry.TABLE_NAME + " (" + ItemEntry._ID + "));";
         
         // create past inventory table
         final String SQL_CREATE_PAST_INVENTORY_TABLE = "CREATE TABLE " +
@@ -76,11 +92,26 @@ public class InventoryDbHelper extends SQLiteOpenHelper {
                 "FOREIGN KEY (" + CurrentInventoryEntry.COLUMN_ITEM_KEY + ") REFERENCES " +
                 ItemEntry.TABLE_NAME + " (" + ItemEntry._ID + "));";
 
+        // create past inventory table
+        final String SQL_CREATE_SHIP_INVENTORY_TABLE = "CREATE TABLE " +
+                ShipInventoryEntry.TABLE_NAME + " (" +
+                ShipInventoryEntry._ID + " INTEGER PRIMARY KEY, " +
+                ShipInventoryEntry.COLUMN_ITEM_KEY + " INTEGER NOT NULL, " +
+                ShipInventoryEntry.COLUMN_QTY + " INTEGER NOT NULL, " +
+                ShipInventoryEntry.COLUMN_DATE_SHIPPED + " INTEGER NOT NULL, " +
+                ShipInventoryEntry.COLUMN_DONOR + " TEXT NOT NULL, " +
+
+                //set up foreign key for item
+                "FOREIGN KEY (" + CurrentInventoryEntry.COLUMN_ITEM_KEY + ") REFERENCES " +
+                ItemEntry.TABLE_NAME + " (" + ItemEntry._ID + "));";
+
         // execute SQL commands
         sqLiteDatabase.execSQL(SQL_CREATE_CATEGORY_TABLE);
         sqLiteDatabase.execSQL(SQL_CREATE_ITEM_TABLE);
         sqLiteDatabase.execSQL(SQL_CREATE_CURRENT_INVENTORY_TABLE);
         sqLiteDatabase.execSQL(SQL_CREATE_PAST_INVENTORY_TABLE);
+        sqLiteDatabase.execSQL(SQL_CREATE_RECEIVE_INVENTORY_TABLE);
+        sqLiteDatabase.execSQL(SQL_CREATE_SHIP_INVENTORY_TABLE);
 
         // insert "Un-categorized" into category table
         ContentValues defaultCategory = new ContentValues();
@@ -92,11 +123,19 @@ public class InventoryDbHelper extends SQLiteOpenHelper {
     @Override
     public void onUpgrade(SQLiteDatabase sqLiteDatabase, int oldVersion, int newVersion) {
         // If different version table, handle changes
-        // For this database revision (4 to 5) drop old tables
-        sqLiteDatabase.execSQL("DROP TABLE IF EXISTS " + CurrentInventoryEntry.TABLE_NAME);
-        sqLiteDatabase.execSQL("DROP TABLE IF EXISTS " + PastInventoryEntry.TABLE_NAME);
-        sqLiteDatabase.execSQL("DROP TABLE IF EXISTS " + ItemEntry.TABLE_NAME);
-        sqLiteDatabase.execSQL("DROP TABLE IF EXISTS " + CategoryEntry.TABLE_NAME);
-        onCreate(sqLiteDatabase);
+
+        // For this database revision (x to 5) drop old tables
+        if (oldVersion < 6) {
+            sqLiteDatabase.execSQL("DROP TABLE IF EXISTS " + CurrentInventoryEntry.TABLE_NAME);
+            sqLiteDatabase.execSQL("DROP TABLE IF EXISTS " + PastInventoryEntry.TABLE_NAME);
+            sqLiteDatabase.execSQL("DROP TABLE IF EXISTS " + ItemEntry.TABLE_NAME);
+            sqLiteDatabase.execSQL("DROP TABLE IF EXISTS " + CategoryEntry.TABLE_NAME);
+            onCreate(sqLiteDatabase);
+        }
+
+        // for database revisions 6
+        if (oldVersion == 6) {
+
+        }
     }
 }
